Change tests to unit instead of integration
