+++ es/amqp/amqp.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/italolelis/outboxer"
-	"github.com/streadway/amqp"
+	amqp "github.com/rabbitmq/amqp091-go"
 )
 
 const (
